Feature: Criado o contexto da aplicação
